refactor(examples): isolate P2P setup
